requirements.txt
fastapi==0.68.1
uvicorn==0.15.0
python-multipart==0.0.5
Pillow==9.3.0
numpy==1.23.5
tensorflow==2.10.0
python-jose==3.3.0
pydantic==1.8.2
matplotlib==3.5.3
scipy==1.9.3
flask==2.2.3
pytesseract==0.3.10
pdf2image==1.16.3
twilio==8.10.0
python-dotenv==1.0.0
werkzeug==2.2.3
gunicorn==20.1.0
phonenumbers==9.0.5
flask-limiter==3.5.0
flask-wtf==1.2.1
bleach==6.1.0
